restructured activity handlers. updated packages.
