+++ BingeBuddyNg/BingeBuddyNg.Services/Activity/Commands/AddDrinkActivityCommand.cs
@@ -1,9 +1,11 @@
 ﻿using BingeBuddyNg.Services.Drink;
+using BingeBuddyNg.Services.Infrastructure.Messaging;
+using BingeBuddyNg.Services.User;
 using BingeBuddyNg.Services.Venue;
 using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace BingeBuddyNg.Services.Activity.Commands
 {
@@ -29,5 +31,40 @@ namespace BingeBuddyNg.Services.Activity.Commands
         public double Volume { get; }
         public Location Location { get; }
         public VenueModel Venue { get; }
+    }
+
+    public class AddDrinkActivityCommandHandler :
+            IRequestHandler<AddDrinkActivityCommand, string>
+    {
+        private readonly IUserRepository userRepository;
+        private readonly IActivityRepository activityRepository;
+        private readonly IMessagingService messagingService;
+
+        public AddDrinkActivityCommandHandler(
+            IUserRepository userRepository,
+            IActivityRepository activityRepository,
+            IMessagingService messagingService)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+            this.activityRepository = activityRepository ?? throw new ArgumentNullException(nameof(activityRepository));
+            this.messagingService = messagingService ?? throw new ArgumentNullException(nameof(messagingService));
+        }
+
+        public async Task<string> Handle(AddDrinkActivityCommand request, CancellationToken cancellationToken)
+        {
+            var user = await this.userRepository.FindUserAsync(request.UserId);
+
+            var activity = Activity.CreateDrinkActivity(DateTime.UtcNow, request.Location, request.UserId, user.Name,
+                request.DrinkType, request.DrinkId, request.DrinkName, request.AlcPrc, request.Volume);
+            activity.Venue = request.Venue;
+
+            var savedActivity = await this.activityRepository.AddActivityAsync(activity);
+
+            await activityRepository.AddToActivityAddedQueueAsync(savedActivity.Id);
+
+            await messagingService.SendMessageAsync(new DrinkEventMessage(request.UserId, request.DrinkId, activity.Timestamp));
+
+            return savedActivity.Id;
+        }
     }
 }
